+++ src/components/layout/HeroSection.jsx
