+++ src/pages/RecruitingManagePage.jsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import AdminHeader from '../components/common/AdminHeader';
-import AdminCodeModal from '../components/common/AdminCodeModal';
-import AdminCodeResultModal from '../components/common/AdminCodeResultModal';
-import AdminManageModal from '../components/common/AdminManageModal';
-import CreateRecruitingModal from '../components/common/CreateRecruitingModal';
-import RecruitingSearchFilter from '../components/recruiting/RecruitingSearchFilter';
-import RecruitingStats from '../components/recruiting/RecruitingStats';
-import RecruitingList from '../components/recruiting/RecruitingList';
-import Pagination from '../components/common/Pagination';
+import AdminHeader from '../features/admin/AdminHeader';
+import AdminCodeModal from '../features/admin/AdminCodeModal';
+import AdminCodeResultModal from '../features/admin/AdminCodeResultModal';
+import AdminManageModal from '../features/admin/AdminManageModal';
+import CreateRecruitingModal from '../features/admin/CreateRecruitingModal';
+import RecruitingSearchFilter from '../features/recruiting/RecruitingSearchFilter';
+import RecruitingStats from '../features/recruiting/RecruitingStats';
+import RecruitingList from '../features/recruiting/RecruitingList';
+import Pagination from '../components/ui/Pagination';
 import { useRecruitingManagement } from '../hooks/useRecruitingManagement';
 import { useModalManagement } from '../hooks/useModalManagement';
 import { ROUTES } from '../constants/routes';
